+++ README.md
@@ -1,6 +1,7 @@
 [![build status](https://secure.travis-ci.org/nlf/riakpbc.png)](http://travis-ci.org/nlf/riakpbc)
 [![NPM version](https://badge.fury.io/js/riakpbc.png)](http://badge.fury.io/js/riakpbc)
 [![Dependency Status](https://gemnasium.com/natural/riakpbc.png)](https://gemnasium.com/natural/riakpbc)
+[![Code Climate](https://codeclimate.com/github/nlf/riakpbc.png)](https://codeclimate.com/github/nlf/riakpbc)
 
 # RiakPBC
 RiakPBC is a low-level [Riak 1.4](http://basho.com/riak)
