+++ index.js
@@ -4,6 +4,7 @@ var net = require('net'),
     butils = require('butils'),
     through = require('through'),
     path = require('path'),
+    _merge = require('./lib/merge'),
     nextTick = setImmediate || process.nextTick;
 
 
@@ -168,28 +169,6 @@ RiakPBC.prototype._processNext = function () {
         });
     }
 };
-
-function _merge(obj1, obj2) {
-    var obj = {};
-    if (obj2.hasOwnProperty('phase')) {
-        obj = obj1;
-        if (obj[obj2.phase] === undefined) obj[obj2.phase] = [];
-        obj[obj2.phase] = obj[obj2.phase].concat(JSON.parse(obj2.response));
-    } else {
-        [obj1, obj2].forEach(function (old) {
-            Object.keys(old).forEach(function (key) {
-                if (!old.hasOwnProperty(key)) return;
-                if (Array.isArray(old[key])) {
-                    if (!obj[key]) obj[key] = [];
-                    obj[key] = obj[key].concat(old[key]);
-                } else {
-                    obj[key] = old[key];
-                }
-            });
-        });
-    }
-    return obj;
-}
 
 RiakPBC.prototype.makeRequest = function (type, data, callback, expectMultiple, streaming) {
     var self = this,
